divide pdfrepor on tools and main part
